refactor: solid principe movie entity
