--- src/domain/domain.module.ts
@@ -1,9 +1,0 @@
-import { HttpModule } from '@/infra/http/http.module';
-import { Module } from '@nestjs/common';
-
-@Module({
-  imports: [HttpModule],
-  controllers: [],
-  providers: [],
-})
-export class DomainModule {}
