+++ src/infra/repositories/database-movies.repository.ts
@@ -3,7 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { MovieEntity } from '../entities/movie.entity';
 import { Movie } from '@/domain/model/movie';
-import { TypeOrmMovieMapper } from '../persistence/typeorm/mapper/typeorm-movie-mapper';
+import { TypeOrmMovieMapper } from '../mapper/typeorm/typeorm-movie-mapper';
 import { MovieRepository } from '@/domain/repositories/movie.repositoy';
 
 @Injectable()
